fix: ios notarization error from nested frameworks (#34)
* Unnest frameworks, create separate framework dependencies

For iOS, BugSplat.xcframework cannot embed other frameworks.

Updated Package.swift to create 3 .binaryTarget one for BugSplat, HockeySDK and CrashReporter

*TODO:*
Each xcframework should be versioned, zipped, and made available for public download from a URL that is published from the respective target URL
The framework version should be used in the download URL.
The checksum for each .zip should also be updated.

*Example Apps:*

BugSplatTest-SwiftUI-SPM was not updated. It should "just work" after the new BugSplat Swift Package is published along with the dependencies. It should be checked after publishing.

The command-line Mac Demo was updated as follows:
BugSplatTest-macOS-Tool-CPlusPlus was updated to include a new Build Phase Script that signs the frameworks after they are copied to the build directory. The Team IDs of the signing in the Build Phase script must match the Team ID for the App (demo).

The following Example_Apps now depend on BugSplat, HockeySDK and CrashReporter Frameworks. All are embedded and signed by the host app.
BugSplatTest-macOS-UIKit-ObjC
BugSplatTest-UIKit-ObjC
BugSplatTest-UIKit-Swift
BugSplatTest-SwiftUI

* chore(build): add HockeySDK and PlCrashReporter to release

* updated README.md

* Apply suggestions from code review

---------

Co-authored-by: Bobby Galli <bobbyg603@pm.me>